app/requirements.txt
pandas
psycopg2-binary
matplotlib
config